requirements.txt
python-telegram-bot[ext]==21.3
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.35
python-dotenv==1.0.1
websockets==12.0
